# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ramon api",
        "description": "api for games, movies and books",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get all books",
                "description": "Returns all books",
                "operationId": "bookIndex",
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/book/{id}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get book by index",
                "description": "Returns a book by the given id",
                "operationId": "showBooks",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Book"
                ],
                "summary": "Updates and returns a book object",
                "description": "Updates and returns a book",
                "operationId": "editBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update a Book object ID not required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Delete a book",
                "description": "Delete a book with the given id",
                "operationId": "destroyBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Books"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/book": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Creates and returns a book object",
                "description": "Creates and returns a book",
                "operationId": "createBook",
                "requestBody": {
                    "description": "Update a Book object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/book/giveTag/{id}": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "add a tag to a book",
                "description": "add tag",
                "operationId": "tagBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Tag object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/game/{id}": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Get game by index",
                "description": "Returns a game by the given id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Edit game",
                "description": "Updates a game Consuming a Game object",
                "operationId": "edit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update a Game object ID not required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Game"
                ],
                "summary": "Delete a game",
                "description": "Delete a game with the given id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    }
}